backend/requirements.txt
fastapi
uvicorn
motor
passlib[bcrypt]
python-jose[cryptography]
python-dotenv
pymongo
